chore: Relocate extension helper file
