+++ e2e/options.test.ts
@@ -1,7 +1,7 @@
 import type { Browser, Page } from 'puppeteer';
 import puppeteer from 'puppeteer';
 import { afterAll, beforeAll, describe, expect, it } from 'vitest';
-import { getExtensionID, getExtensionPath } from '~/utils/chrome-extension';
+import { getExtensionID, getExtensionPath } from '~/lib/chrome/extension';
 
 const extensionPath = getExtensionPath();
 const extensionID = getExtensionID();
